+++ tests/test-meshutilities/meson.build
@@ -1,5 +1,9 @@
-meshutilitiesTest = executable('meshutilitesTest',
-                 ['main.cpp'],
+meshutilitiesTest = executable('test-meshutilites',
+                               ['main.cpp',
+                                'cube.h',
+                                'cube.cpp',
+                                'test_meshutilities.cpp'
+                               ],
                 dependencies: [gtest_dep, gmock_dep, cactus_core_dep])
 
-test('meshutilities Test', meshutilitiesTest)
+test('test-meshutilities', meshutilitiesTest)
